split up tokenization stuff
